telegram-news-analysis/src/api/requirements.txt
Flask==2.1.2
Werkzeug==2.2.2
Flask-Cors==3.0.10
PyJWT==2.3.0
pymongo==4.2.0
elasticsearch==7.14.0
python-dotenv==0.20.0
psycopg2-binary>=2.9.3